monyhar/third_party/blink/web_tests/http/tests/devtools/console/viewport-testing/console-key-links-expected.txt

Tests that console links are keyboard navigable.


Running: testNavigatingLinks
Evaluating: console.log("Before");console.log("Text around www.monyhar.org/1a multiple links, www.monyhar.org/1b");console.log("www.monyhar.org/2");
Message count: 3
Setting focus in prompt:

Shift+Tab:
Viewport virtual selection: 2
activeElement: SPAN.devtools-link
active text: console-key-links.js:26

ArrowUp:
Viewport virtual selection: 2
activeElement: SPAN.devtools-link
active text: www.monyhar.org/2

ArrowUp:
Viewport virtual selection: 2
activeElement: DIV.console-message-wrapper.console-from-api.console-info-level.console-selected
active text: console-key-links.js:26 www.monyhar.org/2

ArrowUp:
Viewport virtual selection: 1
activeElement: SPAN.devtools-link
active text: console-key-links.js:26

ArrowUp:
Viewport virtual selection: 1
activeElement: SPAN.devtools-link
active text: www.monyhar.org/1b

ArrowUp:
Viewport virtual selection: 1
activeElement: SPAN.devtools-link
active text: www.monyhar.org/1a

ArrowUp:
Viewport virtual selection: 1
activeElement: DIV.console-message-wrapper.console-from-api.console-info-level.console-selected
active text: console-key-links.js:26 Text around www.monyhar.org/1a multiple links, www.monyhar.org/1b

ArrowDown:
Viewport virtual selection: 1
activeElement: SPAN.devtools-link
active text: www.monyhar.org/1a

ArrowDown:
Viewport virtual selection: 1
activeElement: SPAN.devtools-link
active text: www.monyhar.org/1b

ArrowDown:
Viewport virtual selection: 1
activeElement: SPAN.devtools-link
active text: console-key-links.js:26

ArrowDown:
Viewport virtual selection: 2
activeElement: DIV.console-message-wrapper.console-from-api.console-info-level.console-selected
active text: console-key-links.js:26 www.monyhar.org/2

ArrowDown:
Viewport virtual selection: 2
activeElement: SPAN.devtools-link
active text: www.monyhar.org/2

ArrowDown:
Viewport virtual selection: 2
activeElement: SPAN.devtools-link
active text: console-key-links.js:26

Running: testNavigatingLinksInStackTrace
Evaluating: fn1()
Message count: 1
Setting focus in prompt:

Shift+Tab:

ArrowUp:
Viewport virtual selection: 0
Is trace expanded: NO
activeElement: SPAN.devtools-link
active text: www.monyhar.org/linkInErrMsg

ArrowDown:

ArrowUp:
Viewport virtual selection: 0
Is trace expanded: NO
activeElement: SPAN.devtools-link
active text: www.monyhar.org/linkInErrMsg

ArrowUp:
Viewport virtual selection: 0
Is trace expanded: NO
activeElement: DIV.console-message-wrapper.console-from-api.console-error-level.console-selected
active text: foo.js:18 Custom error with link www.monyhar.org/linkInErrMsg
fn1 @ foo.js:18
(anonymous) @ console-key-links.js:71

ArrowRight:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: DIV.console-message-wrapper.console-from-api.console-error-level.console-selected
active text: foo.js:18 Custom error with link www.monyhar.org/linkInErrMsg
fn1 @ foo.js:18
(anonymous) @ console-key-links.js:71

ArrowDown:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: SPAN.devtools-link
active text: www.monyhar.org/linkInErrMsg

ArrowDown:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: SPAN.devtools-link
active text: foo.js:18

ArrowDown:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: SPAN.devtools-link
active text: foo.js:18

ArrowDown:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: SPAN.devtools-link
active text: console-key-links.js:71

ArrowDown:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: SPAN.devtools-link
active text: console-key-links.js:71

ArrowLeft:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: DIV.console-message-wrapper.console-from-api.console-error-level.console-selected
active text: foo.js:18 Custom error with link www.monyhar.org/linkInErrMsg
fn1 @ foo.js:18
(anonymous) @ console-key-links.js:71

ArrowLeft:
Viewport virtual selection: 0
Is trace expanded: NO
activeElement: DIV.console-message-wrapper.console-from-api.console-error-level.console-selected
active text: foo.js:18 Custom error with link www.monyhar.org/linkInErrMsg
fn1 @ foo.js:18
(anonymous) @ console-key-links.js:71
